+++ VTK/Graphics/vtkRearrangeFields.cxx
@@ -57,7 +57,8 @@ vtkRearrangeFields::vtkRearrangeFields()
     {
     for (int i = 0; i < vtkDataSetAttributes::NUM_ATTRIBUTES; i++)
       {
-      int l = strlen(vtkDataSetAttributes::GetAttributeTypeAsString(i));
+      int l = static_cast<int>(
+        strlen(vtkDataSetAttributes::GetAttributeTypeAsString(i)));
       for (int c = 0; c < l && c < 10; c++)
         {
         vtkRearrangeFields::AttributeNames[i][c] =
