+++ VTK/Graphics/vtkSplitField.cxx
@@ -52,7 +52,8 @@ vtkSplitField::vtkSplitField()
     {
     for (int i = 0; i < vtkDataSetAttributes::NUM_ATTRIBUTES; i++)
       {
-      int l = strlen(vtkDataSetAttributes::GetAttributeTypeAsString(i));
+      int l = static_cast<int>(
+        strlen(vtkDataSetAttributes::GetAttributeTypeAsString(i)));
       for (int c = 0; c < l && c < 10; c++)
         {
         vtkSplitField::AttributeNames[i][c] =
