+++ VTK/IO/vtkSQLiteQuery.cxx
@@ -118,7 +118,7 @@ bool vtkSQLiteQuery::Execute()
 
   int prepareStatus = vtk_sqlite3_prepare_v2(db, 
                                              this->Query,
-                                             strlen(this->Query),
+                                             static_cast<int>(strlen(this->Query)),
                                              &this->Statement,
                                              &unused_statement);
